Bound recursive fixpoints: divergent programs truncate instead of wedging
Bug 5, root-caused end to end. The two-worker corpus wedge was never
about workers: a recursive MIN whose value grows around a cycle (sssp
with a weighted self-loop on a live source — synthetic corpus data)
DIVERGES under the aggsrc desugar, a documented limitation: the
aggregation runs outside the loop, so the helper fixpoint enumerates
distances 0, 2, 4, ... forever. Workers merely changed how rows split
across epochs; ddmin (with a sound novelty-probe oracle — the first
oracle's probe itself contained the divergent shape) shrank the
trigger to two epochs and five arcs, and a raw stream dump showed
node 1's aggsrc marching past 131071 — TWICE around the old u16
iteration counter.

Three defects fixed:
- Rel::bound_iterations on the Variable feedback path (the classic
  timely `t.inner < max` filter, DEP2_MAX_ITER, default 100k): a
  divergent fixpoint now COMPLETES at the cap with an error naming the
  relation, instead of wedging the worker forever. Dropping values
  past the cap is exact for min/max (dropped values can only be worse)
  and an approximation otherwise. Healthy fixpoints run orders of
  magnitude below the bound.
- Iter u16 -> u32: the old counter silently WRAPPED at 65535,
  violating the Product lattice order with corrupted timestamps. The
  single-worker corpus sweep only ever "passed" sssp because of that
  wrap — honest iteration counting exposed it, hence the sweep now
  pins a small explicit bound.
- The epoch-drive loop honors shutdown and logs a watchdog error after
  10s of non-completion; the post-shutdown drain is bounded (timely's
  own worker thread otherwise spins in step_or_park after the closure
  returns, making join hang — found by stack-sampling the hang).

sssp rejoins the two-worker corpus subset; the previously-#[ignore]d
responsiveness repro now PASSES un-ignored; a new end-to-end test pins
truncation semantics (responsive after divergence, later epochs
processed, node minima exact, clean shutdown). Diagnostics kept:
DEP2_DEBUG_STUCK frontier dumps and DEP2_DEBUG_AGG stream dumps in the
collectors, plus the env-gated ddmin/characterize harnesses.

[engine — cherry-pick candidate for main]

+++ crates/executing/src/lib.rs
@@ -8,4 +8,7 @@ pub mod map;
 pub mod transformer;
 
 pub type Time = reading::Epoch;
-pub type Iter = u16;
+/// Inner (fixpoint) iteration counter. u32: a divergent recursion used to
+/// WRAP the old u16 at 65535, silently corrupting nested timestamps (the
+/// Product lattice order is violated by wraparound).
+pub type Iter = u32;
